Rename plugin for better readability
