+++ packages/pwa-kit-dev/src/configs/webpack/plugins/overrides-resolver-plugin.js
